UserAccountRepository implemented
